// SDKs/twilio.com/twilio_accounts_v1/1.37.3/typescript/src/sdk/models/shared/accountsv1credentialcredentialaws.ts
import { SpeakeasyMetadata, SpeakeasyBase } from "../../../internal/utils";



export class AccountsV1CredentialCredentialAws extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=account_sid" })
  accountSid?: string;

  @SpeakeasyMetadata({ data: "json, name=date_created" })
  dateCreated?: Date;

  @SpeakeasyMetadata({ data: "json, name=date_updated" })
  dateUpdated?: Date;

  @SpeakeasyMetadata({ data: "json, name=friendly_name" })
  friendlyName?: string;

  @SpeakeasyMetadata({ data: "json, name=sid" })
  sid?: string;

  @SpeakeasyMetadata({ data: "json, name=url" })
  url?: string;
}